restructure dsl methods are available in variables definition
